+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.3.0
+
+- Move various roll tables into a single "Random Tables" compendium pack.
+- Replace broken code implementation with a broken roll table.
+- Make broken table a config variable.
+- Make wield power crit/fumble tables and crit/fumble-on into config variables.
+
 # 2.2.0
 
 - Internal code cleanup.